Added new pathfinding algorithms and test project, refactored code, and updated UI
Updated the solution to include a new test project `PathfindingTests.cs` class and some inital tests for pathing algos. Added a new package reference to `Newtonsoft.Json` in `MiningSimulator.csproj` for saving/loading the MiningNodeGrid to a file for testing. Added `test_grid.json` and `test_grid2.json` with new JSON objects for tests.

Changed `getShortestPath` method in `AStarPathfinding.cs` and `DijkstraPathfinding.cs` to add the sorting of interesting items instead of doing that in the `MyExecutePathfinding.cs` class.

Added `HoffmanPavleyPathfinding` and `TravelingSalesmanPathfinding` classes. Updated `MiningNode.cs`, `MiningNodeGrid.cs`, and `Rune.cs` with new methods to allow for cloning for TravelingSalesmanPathfinding.

Added new methods/buttons for saving/loading the grid in `frmMain.Designer.cs`.  Updated initialization of `miningNodeGrid` and added new methods for saving, loading, and printing state.   Added `BitmapJsonConverter` to allow serializing the images for the nodes.